requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.0
nltk==3.8.1 
transformers==4.35.2
sentence-transformers==2.2.2
accelerate==0.28.0
bitsandbytes==0.42.0
huggingface_hub==0.16.4
faiss-cpu==1.7.4
langchain-google-genai==2.0.9
langchain-community==0.3.17
langchain==0.3.18